+++ src/glsl/header.js
@@ -1,10 +1,10 @@
 /**
 *
-*   FILTER.js WebGL Filters
+*   FILTER.js GLSL Filters (WebGL)
 *   @version: @@VERSION@@
 *   @dependencies: Filter.js
 *
-*   JavaScript Image Processing Library (WebGL / SVG Filters)
+*   JavaScript Image Processing Library (WebGL Filters)
 *   https://github.com/foo123/FILTER.js
 *
 **/